Move shared into sdk folder, update package
