--- shared/api/shared.api
@@ -1,19 +1,0 @@
-public final class io/github/composegears/valkyrie/shared/ValkyrieMode : java/lang/Enum {
-	public static final field IconPack Lio/github/composegears/valkyrie/shared/ValkyrieMode;
-	public static final field Simple Lio/github/composegears/valkyrie/shared/ValkyrieMode;
-	public static final field Unspecified Lio/github/composegears/valkyrie/shared/ValkyrieMode;
-	public static fun getEntries ()Lkotlin/enums/EnumEntries;
-	public static fun valueOf (Ljava/lang/String;)Lio/github/composegears/valkyrie/shared/ValkyrieMode;
-	public static fun values ()[Lio/github/composegears/valkyrie/shared/ValkyrieMode;
-}
-
-public final class io/github/composegears/valkyrie/shared/ValkyrieTool : java/lang/Enum {
-	public static final field Editor Lio/github/composegears/valkyrie/shared/ValkyrieTool;
-	public static final field ImageVectorToXml Lio/github/composegears/valkyrie/shared/ValkyrieTool;
-	public static final field SvgToXml Lio/github/composegears/valkyrie/shared/ValkyrieTool;
-	public static final field WebImport Lio/github/composegears/valkyrie/shared/ValkyrieTool;
-	public static fun getEntries ()Lkotlin/enums/EnumEntries;
-	public static fun valueOf (Ljava/lang/String;)Lio/github/composegears/valkyrie/shared/ValkyrieTool;
-	public static fun values ()[Lio/github/composegears/valkyrie/shared/ValkyrieTool;
-}
-
