--- shared/api/shared.klib.api
@@ -1,32 +1,0 @@
-// Klib ABI Dump
-// Targets: [wasmJs]
-// Rendering settings:
-// - Signature version: 2
-// - Show manifest properties: true
-// - Show declarations: true
-
-// Library unique name: <valkyrie:shared>
-final enum class io.github.composegears.valkyrie.shared/ValkyrieMode : kotlin/Enum<io.github.composegears.valkyrie.shared/ValkyrieMode> { // io.github.composegears.valkyrie.shared/ValkyrieMode|null[0]
-    enum entry IconPack // io.github.composegears.valkyrie.shared/ValkyrieMode.IconPack|null[0]
-    enum entry Simple // io.github.composegears.valkyrie.shared/ValkyrieMode.Simple|null[0]
-    enum entry Unspecified // io.github.composegears.valkyrie.shared/ValkyrieMode.Unspecified|null[0]
-
-    final val entries // io.github.composegears.valkyrie.shared/ValkyrieMode.entries|#static{}entries[0]
-        final fun <get-entries>(): kotlin.enums/EnumEntries<io.github.composegears.valkyrie.shared/ValkyrieMode> // io.github.composegears.valkyrie.shared/ValkyrieMode.entries.<get-entries>|<get-entries>#static(){}[0]
-
-    final fun valueOf(kotlin/String): io.github.composegears.valkyrie.shared/ValkyrieMode // io.github.composegears.valkyrie.shared/ValkyrieMode.valueOf|valueOf#static(kotlin.String){}[0]
-    final fun values(): kotlin/Array<io.github.composegears.valkyrie.shared/ValkyrieMode> // io.github.composegears.valkyrie.shared/ValkyrieMode.values|values#static(){}[0]
-}
-
-final enum class io.github.composegears.valkyrie.shared/ValkyrieTool : kotlin/Enum<io.github.composegears.valkyrie.shared/ValkyrieTool> { // io.github.composegears.valkyrie.shared/ValkyrieTool|null[0]
-    enum entry Editor // io.github.composegears.valkyrie.shared/ValkyrieTool.Editor|null[0]
-    enum entry ImageVectorToXml // io.github.composegears.valkyrie.shared/ValkyrieTool.ImageVectorToXml|null[0]
-    enum entry SvgToXml // io.github.composegears.valkyrie.shared/ValkyrieTool.SvgToXml|null[0]
-    enum entry WebImport // io.github.composegears.valkyrie.shared/ValkyrieTool.WebImport|null[0]
-
-    final val entries // io.github.composegears.valkyrie.shared/ValkyrieTool.entries|#static{}entries[0]
-        final fun <get-entries>(): kotlin.enums/EnumEntries<io.github.composegears.valkyrie.shared/ValkyrieTool> // io.github.composegears.valkyrie.shared/ValkyrieTool.entries.<get-entries>|<get-entries>#static(){}[0]
-
-    final fun valueOf(kotlin/String): io.github.composegears.valkyrie.shared/ValkyrieTool // io.github.composegears.valkyrie.shared/ValkyrieTool.valueOf|valueOf#static(kotlin.String){}[0]
-    final fun values(): kotlin/Array<io.github.composegears.valkyrie.shared/ValkyrieTool> // io.github.composegears.valkyrie.shared/ValkyrieTool.values|values#static(){}[0]
-}
