+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/action/ActionUtil.kt
@@ -1,7 +1,7 @@
 package io.github.composegears.valkyrie.action
 
+import io.github.composegears.valkyrie.sdk.shared.ValkyrieMode
 import io.github.composegears.valkyrie.service.PersistentSettings
-import io.github.composegears.valkyrie.shared.ValkyrieMode
 
 val PersistentSettings.ValkyrieState.isIconPackRequired: Boolean
     get() = mode == ValkyrieMode.Unspecified || mode == ValkyrieMode.Simple
