+++ tools/idea-plugin/src/main/kotlin/io/github/composegears/valkyrie/ui/screen/settings/SettingsViewModel.kt
@@ -4,9 +4,9 @@ import androidx.lifecycle.ViewModel
 import com.composegears.leviathan.compose.inject
 import com.intellij.collaboration.async.mapState
 import io.github.composegears.valkyrie.generator.jvm.imagevector.OutputFormat
+import io.github.composegears.valkyrie.sdk.shared.ValkyrieMode
+import io.github.composegears.valkyrie.sdk.shared.ValkyrieMode.Unspecified
 import io.github.composegears.valkyrie.settings.updateOutputFormat
-import io.github.composegears.valkyrie.shared.ValkyrieMode
-import io.github.composegears.valkyrie.shared.ValkyrieMode.Unspecified
 import io.github.composegears.valkyrie.ui.di.DI
 import io.github.composegears.valkyrie.ui.domain.model.PreviewType
 import io.github.composegears.valkyrie.ui.screen.settings.model.SettingsAction
